backend/app/routes/upload.py
import os
import uuid
import logging
from fastapi import APIRouter, UploadFile, File, HTTPException
from pydantic import BaseModel
from typing import Optional
import requests
from PIL import Image
from app.config import HF_API_KEY, HF_VISION_MODEL
from app.db import db

logger = logging.getLogger(__name__)

router = APIRouter()

class UploadResponse(BaseModel):
    image_id: str
    url: str
    label: str
    confidence: float
    meta: dict

@router.post("/api/upload-image", response_model=UploadResponse)
async def upload_image(file: UploadFile = File(...)):
    """Upload and analyze crop image"""
    try:
        # Validate file type
        if not file.content_type or not file.content_type.startswith('image/'):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read file data
        file_data = await file.read()
        
        # Validate file size (10MB limit)
        if len(file_data) > 10 * 1024 * 1024:
            raise HTTPException(status_code=400, detail="File size must be less than 10MB")
        
        # Generate unique filename
        file_extension = os.path.splitext(file.filename)[1] if file.filename else '.jpg'
        unique_filename = f"{uuid.uuid4()}{file_extension}"
        
        # Try to upload to Supabase storage
        storage_url = None
        if db.is_connected():
            try:
                storage_url = db.upload_file("farm-images", unique_filename, file_data)
                logger.info(f"Uploaded to Supabase storage: {storage_url}")
            except Exception as e:
                logger.warning(f"Supabase upload failed: {e}")
        
        # Fallback to local storage
        if not storage_url:
            local_path = f"app/static/{unique_filename}"
            os.makedirs("app/static", exist_ok=True)
            with open(local_path, "wb") as f:
                f.write(file_data)
            storage_url = f"/static/{unique_filename}"
            logger.info(f"Saved locally: {local_path}")
        
        # Analyze image
        label, confidence = await analyze_image(file_data)
        
        # Save image metadata
        image_id = db.insert_image(
            filename=file.filename or unique_filename,
            storage_path=storage_url,
            label=label,
            confidence=confidence
        )
        
        if not image_id:
            image_id = str(uuid.uuid4())
        
        response = UploadResponse(
            image_id=image_id,
            url=storage_url,
            label=label,
            confidence=confidence,
            meta={
                "filename": file.filename,
                "size": len(file_data),
                "storage": "supabase" if db.is_connected() else "local"
            }
        )
        
        logger.info(f"Image processed successfully: {label} (confidence: {confidence})")
        return response
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Image upload failed: {e}")
        raise HTTPException(status_code=500, detail=f"Image upload failed: {str(e)}")

async def analyze_image(image_data: bytes) -> tuple[str, float]:
    """Analyze image using HF Vision API or fallback classification"""
    
    # Try Hugging Face Vision API if available
    if HF_API_KEY and HF_VISION_MODEL:
        try:
            headers = {
                "Authorization": f"Bearer {HF_API_KEY}",
            }
            
            response = requests.post(
                f"https://api-inference.huggingface.co/models/{HF_VISION_MODEL}",
                headers=headers,
                data=image_data,
                timeout=30
            )
            
            if response.status_code == 200:
                result = response.json()
                if isinstance(result, list) and len(result) > 0:
                    top_prediction = result[0]
                    label = top_prediction.get("label", "Unknown crop")
                    confidence = top_prediction.get("score", 0.5)
                    return label, confidence
                    
        except Exception as e:
            logger.warning(f"HF Vision API failed: {e}")
    
    # Fallback to simple image analysis
    try:
        # Basic image validation and classification
        image = Image.open(io.BytesIO(image_data))
        width, height = image.size
        
        # Simple heuristics based on image properties
        if width > 1000 and height > 1000:
            # High resolution suggests field/crop image
            labels = ["Healthy crop", "Wheat field", "Rice field", "Tomato plant", "Cotton crop"]
            confidence = 0.6
        else:
            # Lower resolution might be close-up of leaves/disease
            labels = ["Leaf sample", "Plant disease", "Pest damage", "Nutrient deficiency", "Healthy leaf"]
            confidence = 0.4
        
        # Random selection for demo (in production, use actual ML model)
        import random
        label = random.choice(labels)
        
        return label, confidence
        
    except Exception as e:
        logger.error(f"Image analysis failed: {e}")
        return "Unknown (analysis failed)", 0.2

import io  # Add this import at the top