backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.0.2
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
Pillow==10.1.0
pytest==7.4.3
httpx==0.25.2